Reuse the extractor library via a private Compat.Extractor
Make `.extract` opt-in and keep it out of `com.thoughtworks`:

- `Compat` is now `private[sbtApiMappings]` and exposes a nested `Extractor`
  member, imported with `import Compat.Extractor._`. This avoids leaking the
  `.extract` implicit onto every function via `import Compat._`.
- On Scala 2.12, `Compat.Extractor` is `val Extractor = com.thoughtworks.Extractor`,
  i.e. a re-export of the `com.thoughtworks.extractor` library (added back as a
  Scala-2.12-only dependency). The library is reused as-is; its `sealed`
  `private[thoughtworks]` traits cannot be wrapped, but a plain `val` alias
  re-exports the object so its implicits come in through `import Compat.Extractor._`.
- On Scala 3 `Compat.Extractor` is an `object` whose `extension` turns a
  function/partial function into a `PartialFunction` (already a valid pattern on
  Scala 3). No `com.thoughtworks.Extractor` shim, so the package is not polluted.
