+++ src/main/scala-2.12/com/thoughtworks/sbtApiMappings/Compat.scala
@@ -9,7 +9,7 @@ import sbtcompat.PluginCompat
 /** sbt 1.x specific bindings: only the pieces that genuinely differ from sbt
   * 2.x. Everything that sbt2-compat can express uniformly is plain shared code.
   */
-object Compat {
+private[sbtApiMappings] object Compat {
 
   /** The value type of the `apiMappings` map. */
   type DocUrl = URL
@@ -20,22 +20,9 @@ object Compat {
   def fileToDocKey: Def.Initialize[Task[File => PluginCompat.FileRef]] =
     Def.task { (file: File) => file }
 
-  // On Scala 2.12 a function/partial function cannot be used as a pattern
-  // directly, so `.extract` adapts it into an extractor object.
-
-  final class Extractor[A, B](unlifted: A => Option[B]) {
-    def unapply(a: A): Option[B] = unlifted(a)
-  }
-
-  implicit final class FunctionExtractOps[A, B](private val f: A => Option[B])
-      extends AnyVal {
-    def extract: Extractor[A, B] = new Extractor(f)
-  }
-
-  implicit final class PartialFunctionExtractOps[A, B](
-      private val pf: PartialFunction[A, B]
-  ) extends AnyVal {
-    def extract: Extractor[A, B] = new Extractor(pf.lift)
-  }
-
+  /** The `.extract` pattern support, opted into with
+    * `import Compat.Extractor._`. On Scala 2.12 it simply re-exports the
+    * `com.thoughtworks.extractor` library.
+    */
+  val Extractor = com.thoughtworks.Extractor
 }
