+++ src/main/scala-3/com/thoughtworks/sbtApiMappings/Compat.scala
@@ -10,7 +10,7 @@ import xsbti.HashedVirtualFileRef
 /** sbt 2.x specific bindings: only the pieces that genuinely differ from sbt
   * 1.x. Everything that sbt2-compat can express uniformly is plain shared code.
   */
-object Compat:
+private[sbtApiMappings] object Compat:
 
   /** The value type of the `apiMappings` map. */
   type DocUrl = URI
@@ -27,11 +27,13 @@ object Compat:
       HashedVirtualFileRef.of(file.toPath.toString, ""): PluginCompat.FileRef
     }
 
-  // On Scala 3 a PartialFunction is already a valid pattern, so `.extract` turns
-  // a function (or partial function) into one.
-
-  extension [A, B](f: A => Option[B])
-    def extract: PartialFunction[A, B] = Function.unlift(f)
-
-  extension [A, B](pf: PartialFunction[A, B])
-    def extract: PartialFunction[A, B] = pf
+  /** The `.extract` pattern support, opted into with `import Compat.Extractor._`.
+    * The `com.thoughtworks.extractor` library has no Scala 3 build, but on Scala
+    * 3 a [[scala.PartialFunction]] is already a valid pattern, so a function (or
+    * partial function) only needs to be turned into one.
+    */
+  object Extractor:
+    extension [A, B](f: A => Option[B])
+      def extract: PartialFunction[A, B] = Function.unlift(f)
+    extension [A, B](pf: PartialFunction[A, B])
+      def extract: PartialFunction[A, B] = pf
