+++ src/main/scala/com/thoughtworks/sbtApiMappings/SparkApiMappingRule.scala
@@ -1,7 +1,7 @@
 package com.thoughtworks.sbtApiMappings
 
 import sbt.{AutoPlugin, ModuleID, _}
-import Compat._
+import Compat.Extractor._
 import sbtcompat.PluginCompat
 
 /** @author
